Reorganized frontend files
