+++ RestaurantReview.UI/ClientApp/cypress.json
@@ -5,5 +5,6 @@
     "screenshotsFolder": "cypress/screenshots",
     "videosFolder": "cypress/videos",
     "supportFile": "cypress/support/index.js",
+    
     "baseUrl": "https://localhost:44364/"
 }
