+++ RestaurantReview.UI/ClientApp/src/App.js
@@ -2,13 +2,13 @@ import React, { Component } from 'react';
 import { Route } from 'react-router';
 import Layout from './components/Layout';
 import Home from './components/Home';
-import Login from './components/Login';
-import Categories from './components/Categories';
-import CreateReview from './components/CreateReview';
-import AddRestaurant from './components/AddRestaurant';
-import Register from './components/Register';
-import ShowAllRestaurants from './components/ShowAllRestaurants';
-import RestaurantPage from './components/RestaurantPage';
+import Login from './components/Authentication/Login';
+import Categories from './components/Categories/Categories';
+import CreateReview from './components/Reviews/CreateReview';
+import AddRestaurant from './components/RestaurantPages/AddRestaurant';
+import Register from './components/Authentication/Register';
+import ShowAllRestaurants from './components/RestaurantPages/ShowAllRestaurants';
+import RestaurantPage from './components/RestaurantPages/RestaurantPage';
 import axios from 'axios';
 import './custom.css'
