+++ RestaurantReview.UI/ClientApp/src/components/Authentication/Register.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import "../Css/Register.css";
+import "./Css/Register.css";
 import { Link } from 'react-router-dom';
 import axios from 'axios';
