+++ RestaurantReview.UI/ClientApp/src/components/RestaurantPages/Css/AllRestaurantCards.css
