+++ RestaurantReview.UI/ClientApp/src/components/RestaurantPages/ShowAllRestaurants.js
@@ -1,8 +1,8 @@
 ﻿import React, { Component } from 'react'
 import ReactDOM from 'react-dom';
 import axios from 'axios';
-import { AllRestaurantCard } from './Cards/AllRestaurantCard';
-import '../Css/ShowAllRestaurants.css';
+import { AllRestaurantCard } from './AllRestaurantCard';
+import './Css/ShowAllRestaurants.css';
 import { Container, Row, Col } from 'reactstrap';
