+++ RestaurantReview.UI/RestaurantReview.UI.csproj
@@ -38,6 +38,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="ClientApp\src\components\Profile\Css\" />
+    <Folder Include="ClientApp\src\components\ReactHooks\Css\" />
     <Folder Include="ClientApp\src\Images\" />
   </ItemGroup>
